requirements.txt
absl-py==2.1.0
accelerate==1.0.1
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
albumentations==0.4.3
altair==5.4.1
analytics-python==1.4.post1
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==3.7.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.2.0
av==13.1.0
backoff==1.10.0
bcrypt==4.2.0
bitsandbytes==0.45.0
blinker==1.8.2
blobfile==3.0.0
braceexpand==0.1.7
cachetools==5.5.0
Cartopy==0.23.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
clip-score==0.1.1
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
datasets==2.4.0
diffusers==0.12.1
dill==0.3.5.1
distlib==0.3.9
einops==0.3.0
eva-decord==0.6.1
exceptiongroup==1.2.2
executing==2.1.0
facexlib==0.3.0
fastapi==0.115.4
ffmpy==0.4.0
filelock==3.16.1
filterpy==1.4.5
fire==0.4.0
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
gradio==3.1.4
graphviz==0.20.3
grpcio==1.67.1
h11==0.12.0
httpcore==0.15.0
httpx==0.25.1
huggingface-hub==0.29.3
icecream==2.1.3
identify==2.6.3
idna==3.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
iopath==0.1.10
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kornia==0.6.0
lazy_loader==0.4
linkify-it-py==2.0.3
llvmlite==0.43.0
lmdb==1.5.1
lpips==0.1.4
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
monotonic==1.6
mpi4py
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.13
narwhals==1.11.1
networkx==3.2.1
nodeenv==1.9.1
numba==0.60.0
numpy==1.24.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.1.1
open_clip_torch==2.29.0
openai-clip==1.0.1
opencv-python==4.7.0.72
opencv-python-headless==4.10.0.84
orjson==3.10.10
packaging==24.1
pandas==2.2.3
parameterized==0.9.0
paramiko==3.5.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
pre_commit==4.0.1
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pudb==2019.2
pyarrow==18.0.0
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.18.0
pyiqa==0.1.13
PyNaCl==1.5.0
pyparsing==3.2.0
pyproj==3.6.1
pyshp==2.3.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-multipart==0.0.16
pytorch-lightning==1.4.2
pytorchvideo==0.1.5
pytz==2024.2
PyWavelets==1.6.0
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
responses==0.18.0
rich==13.9.3
rpds-py==0.20.0
safetensors==0.4.5
scikit-image==0.20.0
scikit-learn==1.5.2
scipy==1.9.1
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
starlette==0.41.2
streamlit==1.39.0
sympy==1.13.1
tabulate==0.9.0
taming-transformers==0.0.1
taming-transformers-rom1504==0.0.6
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
test_tube==0.7.5
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==1.0.12
tokenizers==0.15.2
toml==0.10.2
tomli==2.2.1
torch==1.13.1+cu116
torch-fidelity==0.3.0
torchaudio==0.13.1+cu116
torchinfo==1.8.0
torchmetrics==0.11.4
torchvision==0.14.1+cu116
torchviz==0.0.2
tornado==6.4.1
tqdm==4.66.6
transformers==4.37.2
triton==3.1.0
types-regex==2024.9.11.20240912
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
urwid==2.6.16
uvicorn==0.32.0
virtualenv==20.28.0
watchdog==5.0.3
wcwidth==0.2.13
webdataset==0.2.5
Werkzeug==3.0.6
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.17.0
zipp==3.20.2
